set user_friendly_privacy_policy based on old reading level
